# modules/apps/commerce/headless/headless-commerce/headless-commerce-delivery-cart-impl/rest-openapi.yaml
components:
...
    schemas:
...
        CouponCode:
            properties:
                code:
                    type: string
            type: object
        Error:
            properties:
                errorCode:
                    description:
                        Internal error code mapping
                    example: 996
                    readOnly: true
                    type: integer
                errorDescription:
                    example: Unable to find currency. Currency code should be expressed with
                        3-letter ISO 4217 format.
                    readOnly: true
                    type: string
                message:
                    example: No CommerceCurrency exists with the key {groupId=41811, code=US
                        Dollar}
                    readOnly: true
                    type: string
                status:
                    description:
                        HTTP Status code
                    example: 404
                    readOnly: true
                    type: integer
            required:
                - errorCode
                - errorDescription
                - message
                - status
            type: object
...
info:
    description:
        "Headless Delivery Commerce Cart API. A Java client JAR is available for use with the group ID 'com.liferay',
        artifact ID 'com.liferay.headless.commerce.delivery.cart.client', and version '7.0.0'."
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
    title: Headless Delivery Commerce Cart API
    version: v1.0
openapi: 3.0.1
paths:
...
    /carts/{cartId}:
        delete:
            operationId: deleteCart
            parameters:
                - in: path
                  explode: false
                  name: cartId
                  required: true
                  schema:
                      description:
                          Cart Id
                      format: int64
                      type: integer
                  style: simple
            responses:
                204:
                    description:
                        No content
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
            summary: Deletes a Cart by ID.
            tags: ["Cart"]
        get:
            description:
                Retrive information of the given Cart.
            operationId: getCart
            parameters:
                - in: path
                  explode: false
                  name: cartId
                  required: true
                  schema:
                      description:
                          Cart Id
                      format: int64
                      type: integer
                  style: simple
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Cart"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Cart"
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Unexpected error
            tags: ["Cart"]
        patch:
            operationId: patchCart
            parameters:
                - in: path
                  explode: false
                  name: cartId
                  required: true
                  schema:
                      description:
                          Cart Id
                      format: int64
                      type: integer
                  style: simple
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Cart"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Cart"
                required: true
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Cart"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Cart"
                    description:
                        Updated
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Unexpected error
            summary: Updates a Cart.
            tags: ["Cart"]
        put:
            operationId: putCart
            parameters:
                - in: path
                  explode: false
                  name: cartId
                  required: true
                  schema:
                      description:
                          Cart Id
                      format: int64
                      type: integer
                  style: simple
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Cart"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Cart"
                required: true
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Cart"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Cart"
                    description:
                        Updated
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Unexpected error
            summary: Updates a Cart.
            tags: ["Cart"]
...
    /carts/{cartId}/checkout:
        post:
            operationId: postCartCheckout
            parameters:
                - in: path
                  name: cartId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Cart"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Cart"
                    description:
                        ""
            tags: ["Cart"]
...
    /carts/{cartId}/coupon-code:
        post:
            description:
                Add new Items to a Cart, return the whole Cart updated.
            operationId: postCartCouponCode
            parameters:
                - in: path
                  explode: false
                  name: cartId
                  required: true
                  schema:
                      description:
                          Cart Id
                      format: int64
                      type: integer
                  style: simple
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CouponCode"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/CouponCode"
                required: true
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Cart"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Cart"
                    description:
                        Coupon applied to the order
                401:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Error"
                    description:
                        Unexpected error
            tags: ["Cart"]
...
servers:
    - url: /